Remove GenerateWithStratis
